feat: CardLogo 스토리 구현
